requirements.txt
openai>=1.30.0
tweepy==4.14.0
python-dotenv==1.0.0
schedule==1.2.0
pydantic==2.5.0 